# postbox-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Postamat Platform API
  version: 1.0.0-draft.1
  summary: Parcel locker platform for Türkmenpoçta — client app, kiosk terminal, admin panel.
  description: |
    Single source of truth for the HTTP contract between three front-ends
    (Flutter client app, kiosk terminal, admin panel) and the core backend.

    Conventions — see `docs/API-01.md` for prose:

    * All JSON fields are `snake_case`.
    * All identifiers are UUID (v7 preferred).
    * All timestamps are UTC, ISO-8601 with `Z`. Clients render in `Asia/Ashgabat`.
    * All money is an integer in minor units plus a currency code. Never a float.
    * All phone numbers are E.164 (`+993…`).
    * Breaking changes ship as `/api/v2`, never as edits to `/api/v1`.

    Generated clients: `orval`/`openapi-typescript` for web, `openapi-generator` for Dart.
    Do not hand-edit generated code — edit this file.
  contact:
    name: Postamat platform team
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.postamat.tm/api/v1
    description: Production (inside Türkmenistan)
  - url: https://stage.postamat.tm/api/v1
    description: Staging
  - url: http://localhost:4010/api/v1
    description: Prism mock from this file
...
security:
  - clientAuth: []
paths:
...
  /postamats:
    get:
      tags: [Reference]
      operationId: listPostamats
      summary: Postamats for the map and the list
      description: |
        Backs screen `193:1905`: a map with pins plus a list ordered by distance.
        Pass `near_lat`/`near_lon` to get `distance_m` filled in; without them the
        list is ordered by name and `distance_m` is `null` (the app then hides it).
        Blocked postamats are omitted entirely for clients.

        Public — the map, the prices and the free-cell counts are all visible before
        sign-in. Login is required only to actually book.
      security: []
      parameters:
        - name: city_id
          in: query
          schema:
            type: string
            format: uuid
        - name: near_lat
          in: query
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: near_lon
          in: query
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - name: has_free_cells
          in: query
          schema:
            type: boolean
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: Postamats.
          content:
            application/json:
              schema:
                type: object
                required: [items, pagination]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostamatListItem'
                  pagination:
                    $ref: '#/components/schemas/PagePagination'
              example:
                items:
                  - id: 0190f2c1-6b7a-7c3e-9a11-2f0c9a5d4e20
                    name: ТП #4 Беркарарлык
                    address: Почтовый офис 744000, проспект Атамурата Ниязова 72
                    city_id: 0190f2c1-6b7a-7c3e-9a11-2f0c9a5d4e10
                    location:
                      lat: 37.9601
                      lon: 58.3261
                    working_hours: '08:00–20:00'
                    status: active
                    distance_m: 1200
                    free_cell_count: 6
                pagination:
                  page: 1
                  per_page: 20
                  total: 6
                  total_pages: 1
        '400':
          $ref: '#/components/responses/BadRequest'
  /postamats/{postamat_id}:
    get:
      tags: [Reference]
      operationId: getPostamat
      summary: Postamat details with live availability
      description: |
        Backs screen `197:2947` — sizes, dimensions, prices and remaining counts. Public.
      security: []
      parameters:
        - $ref: '#/components/parameters/PostamatId'
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: Postamat.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostamatDetails'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 'Codes: `POSTAMAT_BLOCKED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/postamats:
    get:
      tags: [Admin — postamats]
      operationId: adminListPostamats
      summary: Postamats table
      description: Backs screen `454:3881`, including the occupied/total counter.
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/Query'
        - name: status
          in: query
          description: Comma-separated list of `PostamatStatus`.
          schema:
            type: string
            example: active,maintenance
        - name: city_id
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Postamats.
          content:
            application/json:
              schema:
                type: object
                required: [items, pagination]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminPostamatListItem'
                  pagination:
                    $ref: '#/components/schemas/PagePagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Admin — postamats]
      operationId: adminCreatePostamat
      summary: Create a postamat
      description: |
        Coordinates are picked on the map («Добавить метку на карте»), not typed.
        Cells are created separately — see the bulk endpoint.
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostamatWrite'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostamatDetails'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: |
            Codes: `VALIDATION_ERROR`, `CITY_NOT_AVAILABLE`, `HARDWARE_ADDRESS_DUPLICATE`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/postamats/{postamat_id}:
    get:
      tags: [Admin — postamats]
      operationId: adminGetPostamat
      summary: Postamat with technical data
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/PostamatId'
      responses:
        '200':
          description: Postamat.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPostamatDetails'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Admin — postamats]
      operationId: adminUpdatePostamat
      summary: Edit a postamat
      description: Backs screen `333:4653`. Every field is optional; send what changed.
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/PostamatId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostamatWrite'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPostamatDetails'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /admin/postamats/{postamat_id}/block:
    post:
      tags: [Admin — postamats]
      operationId: adminBlockPostamat
      summary: Block a postamat
      description: |
        This is what the bin icon in the table actually does. There is no `DELETE`
        anywhere in this API: blocking hides the postamat from the client app
        immediately while leaving history, audit log and statistics intact.
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/PostamatId'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 3
                  maxLength: 500
      responses:
        '200':
          description: Blocked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPostamatDetails'
        '409':
          description: |
            Codes: `POSTAMAT_HAS_ACTIVE_BOOKINGS` — with the count in `details`,
            so the panel can ask whether to proceed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/postamats/{postamat_id}/unblock:
    post:
      tags: [Admin — postamats]
      operationId: adminUnblockPostamat
      summary: Unblock a postamat
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/PostamatId'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      responses:
        '200':
          description: Unblocked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPostamatDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/postamats/{postamat_id}/schedule:
    put:
      tags: [Admin — postamats]
      operationId: adminSetPostamatSchedule
      summary: Opening hours
      description: |
        Hours are structured rather than a display string, because expiry depends on
        them. Paying for 12 hours at 19:00 at a site that closes at 20:00 would expire
        at 07:00 — before the recipient could physically arrive. When a postamat is not
        `round_the_clock`, expiry is pushed to shortly after the next opening, and the
        grace period runs from there too.
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/PostamatId'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostamatSchedule'
      responses:
        '200':
          description: Schedule saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostamatSchedule'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: |
            Codes: `VALIDATION_ERROR` — `closes_at` earlier than `opens_at`, or a
            weekday listed twice.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/postamats/{postamat_id}/photos:
    post:
      tags: [Admin — postamats]
      operationId: adminUploadPostamatPhoto
      summary: Upload a photo
      description: 'PNG or JPEG, up to 10 MB — the limit printed in the drop zone.'
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/PostamatId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Uploaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '413':
          description: 'Codes: `FILE_TOO_LARGE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: 'Codes: `UNSUPPORTED_MEDIA_TYPE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/postamats/{postamat_id}/photos/{photo_id}:
    delete:
      tags: [Admin — postamats]
      operationId: adminDeletePostamatPhoto
      summary: Remove a photo
      description: The one genuine delete in this API — a photo carries no history.
      security:
        - adminAuth: []
      parameters:
        - $ref: '#/components/parameters/PostamatId'
        - name: photo_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Removed.
        '404':
          $ref: '#/components/responses/NotFound'

  # ─────────────────────────────────────────────────────── Admin — cell types
...
components:
  securitySchemes:
    clientAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Client access token, 15 minutes. Renew with `/auth/refresh`.
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin access token, 15 minutes. Renew with `/admin/auth/refresh`.
    deviceAuth:
      type: http
      scheme: bearer
      description: |
        Long-lived device token obtained at provisioning. Kept in the device keystore.
    staffAuth:
      type: http
      scheme: bearer
      description: Staff-mode token on a terminal. Short-lived, dies with inactivity.
  parameters:
...
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Mandatory on every state-changing POST. A retry over a flaky connection must
        not book two cells or send two SMS.
      schema:
        type: string
        format: uuid
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: '`tk`, `ru` or `en`. Affects human-readable names, never error codes.'
      schema:
        $ref: '#/components/schemas/Language'
    Page:
      name: page
      in: query
      description: Page number, 1-based. Reference data only.
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
    Sort:
      name: sort
      in: query
      description: 'Field name, `-` prefix for descending. Example: `-created_at`.'
      schema:
        type: string
        example: -created_at
    Query:
      name: q
      in: query
      description: Free-text search. Scope depends on the endpoint.
      schema:
        type: string
        maxLength: 200
    PostamatId:
      name: postamat_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: |
        Malformed request. Codes: `VALIDATION_ERROR`, `IDEMPOTENCY_KEY_MISSING`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: Request body failed validation.
              details:
                fields:
                  phone: must match E.164 (+993…)
              trace_id: 0190f2c1-6b7a-7c3e-9a11-2f0c9a5d4eff
    Unauthorized:
      description: |
        No or expired credentials. Codes: `UNAUTHORIZED`, `TOKEN_EXPIRED`,
        `TOKEN_INVALID`. **401 means refresh the token — never confuse it with 403.**
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: |
        Authenticated but not permitted. Codes: `FORBIDDEN`. **Refreshing will not help.**
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 'Codes: `NOT_FOUND`.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: 'Codes: `VALIDATION_ERROR` and endpoint-specific business rules.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    Language:
      type: string
      enum: [tk, ru, en]
      description: |
        Turkmen is the default: it is preselected on first launch and it is what a
        kiosk falls back to when a session ends. Expressed in prose rather than as
        a schema `default` because code generators emit uncompilable Dart for an
        enum default.
...
    GeoPoint:
      type: object
      required: [lat, lon]
      properties:
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
...
    ErrorResponse:
      type: object
      description: The single error envelope. Every non-2xx response uses this shape.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, trace_id]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
              description: English, for developers and logs. Not shown to end users.
            details:
              type: [object, 'null']
              additionalProperties: true
              description: Structured extras — `attempts_left`, `fields`, `alternatives`.
            trace_id:
              type: string
              description: Correlates with server logs and the audit trail.
      example:
        error:
          code: CELL_ALREADY_TAKEN
          message: This cell was taken by another booking.
          details:
            cell_id: 0190f2c1-6b7a-7c3e-9a11-2f0c9a5d4e70
          trace_id: 0190f2c1-6b7a-7c3e-9a11-2f0c9a5d4eaa

    # ── pagination ────────────────────────────────────────────────────────
    PagePagination:
      type: object
      required: [page, per_page, total, total_pages]
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    PostamatStatus:
      type: string
      enum: [active, maintenance, blocked]
    Postamat:
      type: object
      required: [id, name, address, status]
      properties:
        id:
          type: string
          format: uuid
        number:
          type: string
          description: |
            Human-facing identifier shown in the admin table. Alphanumeric — `10042`
            and `ТП-4` are both valid. Unique across the network.
          example: '10042'
        name:
          type: string
        address:
          type: string
        city_id:
          type: string
          format: uuid
        location:
          $ref: '#/components/schemas/GeoPoint'
        working_hours:
          type: [string, 'null']
          example: '08:00–20:00'
        status:
          $ref: '#/components/schemas/PostamatStatus'
    PostamatListItem:
      allOf:
        - $ref: '#/components/schemas/Postamat'
        - type: object
          properties:
            distance_m:
              type: [integer, 'null']
              description: Null unless `near_lat`/`near_lon` were supplied.
            free_cell_count:
              type: integer
    PostamatDetails:
      allOf:
        - $ref: '#/components/schemas/Postamat'
        - type: object
          properties:
            photos:
              type: array
              items:
                $ref: '#/components/schemas/Photo'
            free_cell_count:
              type: integer
            total_cell_count:
              type: integer
    AdminPostamatListItem:
      allOf:
        - $ref: '#/components/schemas/Postamat'
        - type: object
          required: [occupied_cell_count, total_cell_count]
          properties:
            occupied_cell_count:
              type: integer
            total_cell_count:
              type: integer
            photos:
              type: array
              items:
                $ref: '#/components/schemas/Photo'
    AdminPostamatDetails:
      allOf:
        - $ref: '#/components/schemas/AdminPostamatListItem'
        - type: object
          properties:
            device:
              oneOf:
                - $ref: '#/components/schemas/Device'
                - type: 'null'
            grid_rows:
              type: [integer, 'null']
            grid_cols:
              type: [integer, 'null']
            blocked_reason:
              type: [string, 'null']
            created_at:
              $ref: '#/components/schemas/Timestamp'
            updated_at:
              $ref: '#/components/schemas/Timestamp'
    PostamatWrite:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        number:
          type: string
          maxLength: 32
        address:
          type: string
          maxLength: 500
        city_id:
          type: string
          format: uuid
        location:
          $ref: '#/components/schemas/GeoPoint'
        working_hours:
          type: [string, 'null']
        grid_rows:
          type: [integer, 'null']
          minimum: 1
          maximum: 20
        grid_cols:
          type: [integer, 'null']
          minimum: 1
          maximum: 20
        device:
          type: [object, 'null']
          description: |
            Convenience passthrough — the admin form edits IP and MAC inline even
            though the device is its own resource.
          properties:
            ip_address:
              type: [string, 'null']
              example: 192.168.1.155
            mac_address:
              type: [string, 'null']
              pattern: '^([0-9A-Fa-f]{2}:){5}[0-9A-Fa-f]{2}$'
              example: 00:1A:2B:3C:4D:5E
    PostamatSchedule:
      type: object
      required: [round_the_clock]
      properties:
        round_the_clock:
          type: boolean
          description: |
            True for a cabinet standing in a 24/7 lobby. When true, `days` is ignored.
        days:
          type: array
          description: |
            One entry per weekday the site is open. Missing weekdays mean closed.
            A postamat inside a post office inherits that office's hours.
          items:
            type: object
            required: [weekday, opens_at, closes_at]
            properties:
              weekday:
                type: integer
                minimum: 1
                maximum: 7
                description: 1 = Monday.
              opens_at:
                type: string
                pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
                example: '08:00'
              closes_at:
                type: string
                pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
                example: '20:00'
      example:
        round_the_clock: false
        days:
          - { weekday: 1, opens_at: '08:00', closes_at: '20:00' }
          - { weekday: 2, opens_at: '08:00', closes_at: '20:00' }
...
    Photo:
      type: object
      required: [id, url]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        thumbnail_url:
          type: string
          format: uri
        uploaded_at:
          $ref: '#/components/schemas/Timestamp'
    Device:
      type: object
      description: |
        The Raspberry Pi in the cabinet. Kept as its own resource even though the
        admin screens only surface IP and MAC today.
      required: [id, status]
      properties:
        id:
          type: string
          format: uuid
        postamat_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [online, offline, degraded]
        ip_address:
          type: [string, 'null']
        mac_address:
          type: [string, 'null']
        os_version:
          type: [string, 'null']
        app_version:
          type: [string, 'null']
        last_seen_at:
          type: [string, 'null']
          format: date-time
        certificate_expires_at:
          type: [string, 'null']
          format: date-time
        on_battery:
          type: [boolean, 'null']
          description: Running on UPS.
        tamper_detected:
          type: [boolean, 'null']
        jammed_cell_ids:
          type: array
          items:
            type: string
            format: uuid

    # ── cells ─────────────────────────────────────────────────────────────
...